livekit-pipecat-demo/agent/requirements.txt
pipecat-ai[livekit,silero]
livekit-api